Refactor overlap cases in RangeSet
